TT-28: Create player view
